feat: useFileSelectOptions
