+++ src/components/organisms/FileCompareModal/FileCompareModal.tsx
@@ -1,16 +1,18 @@
-import {useCallback, useEffect, useMemo, useState} from 'react';
+import {useEffect, useMemo, useState} from 'react';
 import {MonacoDiffEditor, monaco} from 'react-monaco-editor';
 import {useWindowSize} from 'react-use';
 
 import {Modal, Select} from 'antd';
 
 import fs from 'fs';
-import {join, sep} from 'path';
+import {join} from 'path';
 
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
 import {closeFileCompareModal} from '@redux/reducers/ui';
 
 import {SelectItemImage} from '@components/atoms';
+
+import {useFileSelectOptions} from '@hooks/useFileSelectOptions';
 
 import {KUBESHOP_MONACO_THEME} from '@utils/monaco';
 
@@ -40,7 +42,9 @@ const FileCompareModal: React.FC = () => {
 
   const {height, width} = useWindowSize();
 
-  const rootFilePath = useMemo(() => fileMap[ROOT_FILE_ENTRY].filePath, []);
+  const rootFilePath = useMemo(() => fileMap[ROOT_FILE_ENTRY].filePath, [fileMap]);
+
+  const fileSelectOptions = useFileSelectOptions();
 
   const comparingFileCode = useMemo(() => {
     if (!comparingFilePath || !rootFilePath) {
@@ -50,27 +54,6 @@ const FileCompareModal: React.FC = () => {
     const absoluteFilePath = join(rootFilePath, comparingFilePath);
     return fs.readFileSync(absoluteFilePath, 'utf-8');
   }, [comparingFilePath, rootFilePath]);
-
-  const filesList: string[] = useMemo(() => {
-    const files: string[] = [];
-
-    Object.entries(fileMap).forEach(([key, value]) => {
-      if (value.children || !value.isSupported || value.isExcluded) {
-        return;
-      }
-      files.push(key.replace(sep, ''));
-    });
-
-    return files;
-  }, [fileMap]);
-
-  const renderFileSelectOptions = useCallback(() => {
-    return filesList.map(fileName => (
-      <Select.Option key={fileName} value={fileName}>
-        {fileName}
-      </Select.Option>
-    ));
-  }, [filesList]);
 
   useEffect(() => {
     if (!currentFilePath || !fileMap[currentFilePath]) {
@@ -99,18 +82,18 @@ const FileCompareModal: React.FC = () => {
       centered
       title={
         <S.TitleContainer>
-          <S.Title $width={(width * SIZE_PERCENTAGE - 100) / 2}>
+          <S.Title $width={(width * SIZE_PERCENTAGE - 175) / 2}>
             Compare <S.TitleFilePath>{currentFilePath}</S.TitleFilePath>
           </S.Title>
-
+          to
           <Select
-            style={{width: '400px', justifySelf: 'flex-end'}}
+            style={{paddingLeft: '50px'}}
             showSearch
             onChange={setComparingFilePath}
             value={comparingFilePath}
             placeholder="Select a comparing file"
           >
-            {renderFileSelectOptions()}
+            {fileSelectOptions}
           </Select>
         </S.TitleContainer>
       }
